Exclude HelpfulFunctions
